# rpc/rest/zmq: reduce LOCK(cs_main) scope: ~6 times as many requests per second