@fanquake 2022-10-14
https://github.com/bitcoin/bitcoin/pull/26308/checks?check_run_id=8878540627
https://github.com/bitcoin/bitcoin/pull/26308/checks?check_run_id=8878540635
```bash
  CXX      rpc/libbitcoin_node_a-node.o
policy/fees.cpp:1042:10: error: calling function 'MaybeCheckNotHeld' requires negative capability '!m_insecure_rand_mutex' [-Werror,-Wthread-safety-analysis]
         WITH_LOCK(m_insecure_rand_mutex, return insecure_rand.rand32()) % 3 != 0)) {
         ^
./sync.h:302:30: note: expanded from macro 'WITH_LOCK'
#define WITH_LOCK(cs, code) (MaybeCheckNotHeld(cs), [&]() -> decltype(auto) { LOCK(cs); code; }())
                             ^
1 error generated.
```