@andrewtoth 2022-10-14
After reading this comment https://github.com/bitcoin/bitcoin/issues/17161#issuecomment-980043633 it appears it is unsafe to call `ReadBlockFromDisk` without holding `cs_main`. The block could be removed while the file is open which can cause issues on Windows.