@andrewtoth 2022-11-18
>Or is the concern not that this change would cause RPC and rest calls to do the wrong thing, but that this change would cause unlink calls to fail on windows and result in a wasted disk space? If that is the bug, and I think a more minimal and elegant fix might just be to make the pruning code resilient to files temporarily being unlinkable because they are in-use, and to just retry and delete them later on the next round of pruning.

@ryanofsky yes, thank you for the suggestion. I've implemented that in #26533. I think that's a better/less invasive solution than what I have proposed in #26316.